requirements.txt
openai
langchain
python-dotenv
chromadb
pypdf
flask